tests/wpt/web-platform-tests/css/css-anchor-position/anchor-position-multicol-colspan-002.html

<!DOCTYPE html>
<title>Anchors in column-spanner in multicol</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#propdef-anchor-name">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-pos">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#anchor-size">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
.relpos {
  position: relative;
}
.columns {
  column-count: 2;
  column-fill: auto;
  column-gap: 10px;
  column-width: 100px;
  width: 210px;
  height: 100px;
}
.colspan {
  column-span: all;
}
.spacer {
  height: 10px;
  background: pink;
}
.anchor1 {
  anchor-name: --a1;
  margin-left: 10px;
  width: 40px;
  background: blue;
}
.anchor1 .spacer {
  background: inherit;
}
.target {
  position: absolute;
  left: anchor(--a1 left);
  top: anchor(--a1 top);
  width: anchor-size(--a1 width);
  height: anchor-size(--a1 height);
  background: lime;
  opacity: .3;
}
</style>
<body onload="checkLayoutForAnchorPos('.target')">
  <div class="relpos">
    <div class="spacer" style="height: 10px"></div>
    <div class="columns relpos">
      <div class="spacer" style="height: 50px"></div>
      <div class="relpos">
        <div class="anchor1">
          <div class="spacer" style="height: 10px"></div>
          <div class="colspan" style="height: 20px"></div>
          <div class="spacer" style="height: 30px"></div>
        </div>
        <!--
          When anchors are split by the column spanners, OOFs laid out in a
          multicol (i.e., the containing block is within a multicol) can't
          compute the united bounding box, because the coordinate system in the
          multicol can't express united rectangles across columns that are not
          lined up.
          <div class="target"></div>
        -->
      </div>
      <div class="target"
           data-offset-x=10 data-offset-y=20
           data-expected-width=150 data-expected-height=60></div>
    </div>
    <div class="target"
          data-offset-x=10 data-offset-y=30
          data-expected-width=150 data-expected-height=60></div>
  </div>
</body>
