LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-multicol-colspan-002-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-offset-x="10" data-offset-y="20" data-expected-width="150" data-expected-height="60"></div>
width expected 150 but got 0
FAIL .target 2 assert_equals:
<div class="target" data-offset-x="10" data-offset-y="30" data-expected-width="150" data-expected-height="60"></div>
width expected 150 but got 0
